More post-merge fixes
